// stable-diffusion/generation/generation_pb.d.ts
export class Prompt extends jspb.Message {
  hasParameters(): boolean;
  clearParameters(): void;
  getParameters(): PromptParameters | undefined;
  setParameters(value?: PromptParameters): void;

  hasText(): boolean;
  clearText(): void;
  getText(): string;
  setText(value: string): void;

  hasTokens(): boolean;
  clearTokens(): void;
  getTokens(): Tokens | undefined;
  setTokens(value?: Tokens): void;

  hasArtifact(): boolean;
  clearArtifact(): void;
  getArtifact(): Artifact | undefined;
  setArtifact(value?: Artifact): void;

  getPromptCase(): Prompt.PromptCase;
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Prompt.AsObject;
  static toObject(includeInstance: boolean, msg: Prompt): Prompt.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: Prompt, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Prompt;
  static deserializeBinaryFromReader(message: Prompt, reader: jspb.BinaryReader): Prompt;
}

export namespace Prompt {
  export type AsObject = {
    parameters?: PromptParameters.AsObject,
    text: string,
    tokens?: Tokens.AsObject,
    artifact?: Artifact.AsObject,
  }

  export enum PromptCase {
    PROMPT_NOT_SET = 0,
    TEXT = 2,
    TOKENS = 3,
    ARTIFACT = 4,
  }
}

export class SamplerParameters extends jspb.Message {
  hasEta(): boolean;
  clearEta(): void;
  getEta(): number;
  setEta(value: number): void;

  hasSamplingSteps(): boolean;
  clearSamplingSteps(): void;
  getSamplingSteps(): number;
  setSamplingSteps(value: number): void;

  hasLatentChannels(): boolean;
  clearLatentChannels(): void;
  getLatentChannels(): number;
  setLatentChannels(value: number): void;

  hasDownsamplingFactor(): boolean;
  clearDownsamplingFactor(): void;
  getDownsamplingFactor(): number;
  setDownsamplingFactor(value: number): void;

  hasCfgScale(): boolean;
  clearCfgScale(): void;
  getCfgScale(): number;
  setCfgScale(value: number): void;

  hasInitNoiseScale(): boolean;
  clearInitNoiseScale(): void;
  getInitNoiseScale(): number;
  setInitNoiseScale(value: number): void;

  hasStepNoiseScale(): boolean;
  clearStepNoiseScale(): void;
  getStepNoiseScale(): number;
  setStepNoiseScale(value: number): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): SamplerParameters.AsObject;
  static toObject(includeInstance: boolean, msg: SamplerParameters): SamplerParameters.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: SamplerParameters, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): SamplerParameters;
  static deserializeBinaryFromReader(message: SamplerParameters, reader: jspb.BinaryReader): SamplerParameters;
}

export namespace SamplerParameters {
  export type AsObject = {
    eta: number,
    samplingSteps: number,
    latentChannels: number,
    downsamplingFactor: number,
    cfgScale: number,
    initNoiseScale: number,
    stepNoiseScale: number,
  }
}
// ...
export class StepParameter extends jspb.Message {
  getScaledStep(): number;
  setScaledStep(value: number): void;

  hasSampler(): boolean;
  clearSampler(): void;
  getSampler(): SamplerParameters | undefined;
  setSampler(value?: SamplerParameters): void;

  hasSchedule(): boolean;
  clearSchedule(): void;
  getSchedule(): ScheduleParameters | undefined;
  setSchedule(value?: ScheduleParameters): void;

  hasGuidance(): boolean;
  clearGuidance(): void;
  getGuidance(): GuidanceParameters | undefined;
  setGuidance(value?: GuidanceParameters): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): StepParameter.AsObject;
  static toObject(includeInstance: boolean, msg: StepParameter): StepParameter.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: StepParameter, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): StepParameter;
  static deserializeBinaryFromReader(message: StepParameter, reader: jspb.BinaryReader): StepParameter;
}

export namespace StepParameter {
  export type AsObject = {
    scaledStep: number,
    sampler?: SamplerParameters.AsObject,
    schedule?: ScheduleParameters.AsObject,
    guidance?: GuidanceParameters.AsObject,
  }
}
// ...
export class TransformType extends jspb.Message {
  hasDiffusion(): boolean;
  clearDiffusion(): void;
  getDiffusion(): DiffusionSamplerMap[keyof DiffusionSamplerMap];
  setDiffusion(value: DiffusionSamplerMap[keyof DiffusionSamplerMap]): void;

  hasUpscaler(): boolean;
  clearUpscaler(): void;
  getUpscaler(): UpscalerMap[keyof UpscalerMap];
  setUpscaler(value: UpscalerMap[keyof UpscalerMap]): void;

  getTypeCase(): TransformType.TypeCase;
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): TransformType.AsObject;
  static toObject(includeInstance: boolean, msg: TransformType): TransformType.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: TransformType, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): TransformType;
  static deserializeBinaryFromReader(message: TransformType, reader: jspb.BinaryReader): TransformType;
}

export namespace TransformType {
  export type AsObject = {
    diffusion: DiffusionSamplerMap[keyof DiffusionSamplerMap],
    upscaler: UpscalerMap[keyof UpscalerMap],
  }

  export enum TypeCase {
    TYPE_NOT_SET = 0,
    DIFFUSION = 1,
    UPSCALER = 2,
  }
}

export class ImageParameters extends jspb.Message {
  hasHeight(): boolean;
  clearHeight(): void;
  getHeight(): number;
  setHeight(value: number): void;

  hasWidth(): boolean;
  clearWidth(): void;
  getWidth(): number;
  setWidth(value: number): void;

  clearSeedList(): void;
  getSeedList(): Array<number>;
  setSeedList(value: Array<number>): void;
  addSeed(value: number, index?: number): number;

  hasSamples(): boolean;
  clearSamples(): void;
  getSamples(): number;
  setSamples(value: number): void;

  hasSteps(): boolean;
  clearSteps(): void;
  getSteps(): number;
  setSteps(value: number): void;

  hasTransform(): boolean;
  clearTransform(): void;
  getTransform(): TransformType | undefined;
  setTransform(value?: TransformType): void;

  clearParametersList(): void;
  getParametersList(): Array<StepParameter>;
  setParametersList(value: Array<StepParameter>): void;
  addParameters(value?: StepParameter, index?: number): StepParameter;

  hasMaskedAreaInit(): boolean;
  clearMaskedAreaInit(): void;
  getMaskedAreaInit(): MaskedAreaInitMap[keyof MaskedAreaInitMap];
  setMaskedAreaInit(value: MaskedAreaInitMap[keyof MaskedAreaInitMap]): void;

  hasWeightMethod(): boolean;
  clearWeightMethod(): void;
  getWeightMethod(): WeightMethodMap[keyof WeightMethodMap];
  setWeightMethod(value: WeightMethodMap[keyof WeightMethodMap]): void;

  hasQuantize(): boolean;
  clearQuantize(): void;
  getQuantize(): boolean;
  setQuantize(value: boolean): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ImageParameters.AsObject;
  static toObject(includeInstance: boolean, msg: ImageParameters): ImageParameters.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: ImageParameters, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ImageParameters;
  static deserializeBinaryFromReader(message: ImageParameters, reader: jspb.BinaryReader): ImageParameters;
}

export namespace ImageParameters {
  export type AsObject = {
    height: number,
    width: number,
    seedList: Array<number>,
    samples: number,
    steps: number,
    transform?: TransformType.AsObject,
    parametersList: Array<StepParameter.AsObject>,
    maskedAreaInit: MaskedAreaInitMap[keyof MaskedAreaInitMap],
    weightMethod: WeightMethodMap[keyof WeightMethodMap],
    quantize: boolean,
  }
}
// ...
export class Request extends jspb.Message {
  getEngineId(): string;
  setEngineId(value: string): void;

  getRequestId(): string;
  setRequestId(value: string): void;

  getRequestedType(): ArtifactTypeMap[keyof ArtifactTypeMap];
  setRequestedType(value: ArtifactTypeMap[keyof ArtifactTypeMap]): void;

  clearPromptList(): void;
  getPromptList(): Array<Prompt>;
  setPromptList(value: Array<Prompt>): void;
  addPrompt(value?: Prompt, index?: number): Prompt;

  hasImage(): boolean;
  clearImage(): void;
  getImage(): ImageParameters | undefined;
  setImage(value?: ImageParameters): void;

  hasClassifier(): boolean;
  clearClassifier(): void;
  getClassifier(): ClassifierParameters | undefined;
  setClassifier(value?: ClassifierParameters): void;

  hasAsset(): boolean;
  clearAsset(): void;
  getAsset(): AssetParameters | undefined;
  setAsset(value?: AssetParameters): void;

  hasInterpolate(): boolean;
  clearInterpolate(): void;
  getInterpolate(): InterpolateParameters | undefined;
  setInterpolate(value?: InterpolateParameters): void;

  hasTransform(): boolean;
  clearTransform(): void;
  getTransform(): TransformParameters | undefined;
  setTransform(value?: TransformParameters): void;

  hasConditioner(): boolean;
  clearConditioner(): void;
  getConditioner(): ConditionerParameters | undefined;
  setConditioner(value?: ConditionerParameters): void;

  hasExtras(): boolean;
  clearExtras(): void;
  getExtras(): google_protobuf_struct_pb.Struct | undefined;
  setExtras(value?: google_protobuf_struct_pb.Struct): void;

  getParamsCase(): Request.ParamsCase;
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Request.AsObject;
  static toObject(includeInstance: boolean, msg: Request): Request.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: Request, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Request;
  static deserializeBinaryFromReader(message: Request, reader: jspb.BinaryReader): Request;
}

export namespace Request {
  export type AsObject = {
    engineId: string,
    requestId: string,
    requestedType: ArtifactTypeMap[keyof ArtifactTypeMap],
    promptList: Array<Prompt.AsObject>,
    image?: ImageParameters.AsObject,
    classifier?: ClassifierParameters.AsObject,
    asset?: AssetParameters.AsObject,
    interpolate?: InterpolateParameters.AsObject,
    transform?: TransformParameters.AsObject,
    conditioner?: ConditionerParameters.AsObject,
    extras?: google_protobuf_struct_pb.Struct.AsObject,
  }

  export enum ParamsCase {
    PARAMS_NOT_SET = 0,
    IMAGE = 5,
    CLASSIFIER = 7,
    ASSET = 8,
    INTERPOLATE = 11,
    TRANSFORM = 12,
  }
}
// ...
export interface DiffusionSamplerMap {
  SAMPLER_DDIM: 0;
  SAMPLER_DDPM: 1;
  SAMPLER_K_EULER: 2;
  SAMPLER_K_EULER_ANCESTRAL: 3;
  SAMPLER_K_HEUN: 4;
  SAMPLER_K_DPM_2: 5;
  SAMPLER_K_DPM_2_ANCESTRAL: 6;
  SAMPLER_K_LMS: 7;
  SAMPLER_K_DPMPP_2S_ANCESTRAL: 8;
  SAMPLER_K_DPMPP_2M: 9;
  SAMPLER_K_DPMPP_SDE: 10;
}

export const DiffusionSampler: DiffusionSamplerMap;

export interface UpscalerMap {
  UPSCALER_RGB: 0;
  UPSCALER_GFPGAN: 1;
  UPSCALER_ESRGAN: 2;
}

export const Upscaler: UpscalerMap;

export interface GuidancePresetMap {
  GUIDANCE_PRESET_NONE: 0;
  GUIDANCE_PRESET_SIMPLE: 1;
  GUIDANCE_PRESET_FAST_BLUE: 2;
  GUIDANCE_PRESET_FAST_GREEN: 3;
  GUIDANCE_PRESET_SLOW: 4;
  GUIDANCE_PRESET_SLOWER: 5;
  GUIDANCE_PRESET_SLOWEST: 6;
}